Server/requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.2.1
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
future==0.18.2
gkeepapi==0.13.6
gpsoauth==0.4.1
html5lib==1.0.1
idna==2.8
importlib-metadata==2.0.0
ipaddr==2.2.0
itsdangerous==1.1.0
jeepney==0.6.0
Jinja2==2.11.2
keyring==21.5.0
keyrings.alt==4.0.1
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pycparser==2.20
pycryptodomex==3.9.9
pyparsing==2.4.6
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
SecretStorage==3.2.0
six==1.14.0
urllib3==1.25.8
waitress==1.4.4
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.4.0
